refactor : member 파일로 관련 엔티티 이동
