+++ backend/src/main/java/com/example/backend/entity/member/Authority.java
@@ -1,4 +1,4 @@
-package com.example.backend.entity;
+package com.example.backend.entity.member;
 
 import jakarta.persistence.*;
 import lombok.Getter;
