+++ backend/src/main/java/com/example/backend/entity/member/EAuthority.java
@@ -1,4 +1,4 @@
-package com.example.backend.entity;
+package com.example.backend.entity.member;
 
 public enum EAuthority {
     ROLE_USER, ROLE_ADMIN
