+++ backend/src/main/java/com/example/backend/entity/member/OAuthProvider.java
@@ -1,4 +1,4 @@
-package com.example.backend.entity;
+package com.example.backend.entity.member;
 
 public enum OAuthProvider {
     NAVER, GOOGLE, GITHUB
